graphic: fix exception due to float being used as param of QColor
  TypeError: arguments did not match any overloaded call:
    QColor(color: Qt.GlobalColor): argument 1 has unexpected type 'int'
    QColor(rgb: int): too many arguments
    QColor(rgba64: QRgba64): argument 1 has unexpected type 'int'
    QColor(variant: Any): too many arguments
    QColor(): too many arguments
    QColor(r: int, g: int, b: int, alpha: int = 255): argument 2 has unexpected type 'float'
    QColor(aname: Optional[str]): argument 1 has unexpected type 'int'
    QColor(acolor: Union[QColor, Qt.GlobalColor]): argument 1 has unexpected type 'int'

An exception is thrown because the green value passed to QColor is a float.
Convert the value to an int before passing it to QColor.

Fixes: https://github.com/oso/qgis-etri/issues/9

+++ qgis_etri/graphic.py
@@ -210,7 +210,7 @@ class QGraphicsSceneEtri(QGraphicsScene):
 
     def __get_category_color(self, i):
         n = len(self.model.categories)
-        g = 255 - 220 * (n - i) / n
+        g = int(255 - 220 * (n - i) / n)
         return QColor(0, g, 0)
 
     def __create_category(self, i, path_below, path_above):
